+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## `0.7.0`
 
+- `shp` 模块：
+  - 新增 GeoJSON 相关的函数：`geometry_to_dict`、`get_geojson_geometries`、`get_geojson_properties`、`make_feature`、`make_geojson`。
+  - 新增 Shapefile 相关的函数：`geometry_to_shape`。
+  - 新增调整多边形绕行方向的函数 `orient_polygon`。
+  - `polygon_to_mask` 函数改名为 `polygon_mask`，修正点落在边界上的行为，新增 `include_boundary` 参数，优化效率。同时新增二维直线网格专用的 `polygon_mask2` 函数。
 - `calc` 模块：
   - `lon_lat_to_xyz` 函数改名为 `lonlat_to_xyz`。
   - `hms_to_degrees` 函数改名为 `dms_to_dd`；移除 `hms_to_degrees2` 函数。